# board_login/swagger.json
{
  "swagger": "2.0",
  "schemes": [
    "http"
  ],
  "host": "localhost:3009",
  "basePath": "/",
  "info":
  {
    "description": "UNI SOPT 7차 개발 합동세미나 API Reference 입니다. <!-- ReDoc-Inject: <security-definitions> -->\n",
    "version": "1.0.0",
    "title": "UNI SOPT _ SERVER PART",
    "x-logo":
    {
      "url": "https://s3.ap-northeast-2.amazonaws.com/sopt21th/sopt21th_logo.png"
    }
  },
  "tags": [
  {
    "name": "sign",
    "description": "로그인, 회원가입 API"
  },
  {
    "name": "board",
    "description": "게시판 API"
  },
  {
    "name": "mypage",
    "description": "마이페이지 API"
  }],
  "x-tagGroups": [
  {
    "name": "sign page",
    "tags": [
      "sign"
    ],
    "name": "board page",
    "tags": [
      "board"
    ],
    "name": "mypage page",
    "tags": [
      "mypage"
    ]
  }],
  "securityDefinitions":
  {
    "petstore_auth":
    {
      "description": "Get access to data while protecting your account credentials.\nOAuth2 is also a safer and more secure way to give you access.\n",
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/api/oauth/dialog",
      "flow": "implicit",
      "scopes":
      {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key":
    {
      "description": "For this sample, you can use the api key `special-key` to test the authorization filters.\n",
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "paths":
  {
    "/signin":
    {
      "post":
      {
        "tags": [
          "sign"
        ],
        "summary": "로그인",
        "description": "email / pwd로 로그인합니다. Request : application/x-www-form-urlencoded",
        "operationId": "sign in",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
        {
          "in": "body",
          "name": "body",
          "description": "유저 로그인을 하기 위한 email, pwd 전달",
          "required": true,
          "schema":
          {
            "type": "object",
            "properties":
            {
              "email":
              {
                "description": "유저 email",
                "type": "string",
                "minLength": 1,
                "required": true,
                "example": "test@sopt.org"
              },
              "pwd":
              {
                "description": "로그인에 사용할 비밀번호",
                "type": "string",
                "minLength": 1,
                "required": true,
                "example": "qwer1234"
              }
            }
          }
        }],
        "responses":
        {
          "201":
          {
            "description": "로그인 성공",
            "schema":
            {
              "type": "object",
              "properties":
              {
                "status":
                {
                  "type": "string",
                  "description": "성공 알림",
                  "example": "success"
                },
                "data":
                {
                  "type": "object",
                  "properties":
                  {
                    "email":
                    {
                      "description": "사용자 email",
                      "type": "string",
                      "example": "test@sopt.org"
                    },
                    "nickname":
                    {
                      "description": "이용자가 등록한 닉네임",
                      "type": "string",
                      "example": "테스트"
                    },
                    "profile":
                    {
                      "description": "이용자의 프로필사진 url",
                      "type": "string",
                      "example": "profile image url"
                    }
                  }

                },
                "msg":
                {
                  "type": "string",
                  "description": "로그인 성공 알림메세지",
                  "example": "successful signin"
                }
              }
            }
          },
          "401":
          {
            "description": "find user data error // signin fail",
            "schema":
            {
              "$ref": "#/definitions/failMessage"
            }
          },
          "500":
          {
            "description": "get connection error // find user data error",
            "schema":
            {
              "$ref": "#/definitions/failMessage"
            }
          }
        }
      }
    },
    "/signup":
    {
      "post":
      {
        "tags": [
          "sign"
        ],
        "summary": "회원가입",
        "description": "사용자 회원가입을 합니다. Request : multipart/form-data",
        "operationId": "sign up",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
        {
          "in": "body",
          "name": "body",
          "description": "유저 회원가입을 하기 위한 email, pwd, nickname, profileImage 전달",
          "required": true,
          "schema":
          {
            "type": "object",
            "properties":
            {
              "email":
              {
                "description": "유저 email",
                "type": "string",
                "minLength": 1,
                "required": true,
                "example": "test@sopt.org"
              },
              "pwd":
              {
                "description": "비밀번호",
                "type": "string",
                "minLength": 1,
                "required": true,
                "example": "qwer1234"
              },
              "nickname":
              {
                "description": "유저 닉네임",
                "type": "string",
                "minLength": 1,
                "required": true,
                "example": "테스트"
              },
              "image":
              {
                "description": "유저의 프로필로 등록할 이미지파일",
                "required": true,
                "type": "image"
              }
            }
          }
        }],
        "responses":
        {
          "201":
          {
            "description": "이용자 회원가입 성공",
            "schema":
            {
              "type": "object",
              "properties":
              {
                "status":
                {
                  "type": "string",
                  "description": "성공 알림",
                  "example": "success"
                },
                "msg":
                {
                  "type": "string",
                  "description": "회원가입 성공 알림 메세지",
                  "example": "successful signup"
                }
              }
            }
          },
          "401":
          {
            "description": "email overlap",
            "schema":
            {
              "$ref": "#/definitions/failMessage"
            }
          },
          "500":
          {
            "description": "get connection error // get user data error // hash password error // insert user data error",
            "schema":
            {
              "$ref": "#/definitions/failMessage"
            }
          }
        }
      }
    },
    "/board/list":
    {
      "get":
      {
        "tags": [
          "board"
        ],
        "summary": "글 리스트 조회",
        "description": "게시판의 글 리스트를 조회합니다. Request : application/x-www-form-urlencoded",
        "operationId": "basicInfo",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [],
        "responses":
        {
          "200":
          {
            "description": "글 리스트",
            "schema":
            {
              "type": "object",
              "properties":
              {
                "status":
                {
                  "type": "string",
                  "description": "성공 알림",
                  "example": "success"
                },
                "data":
                {
                  "type": "array",
                  "items":
                  {
                    "type": "object",
                    "properties":
                    {
                      "board_id":
                      {
                        "type": "integer",
                        "description": "글의 ID",
                        "example": 1
                      },
                      "board_title":
                      {
                        "type": "string",
                        "description": "글 제목",
                        "example": "글쓰기 테스트"
                      },
                      "board_image":
                      {
                        "type": "string",
                        "description": "글 이미지 URL",
                        "example": "image url"
                      },
                      "board_hits":
                      {
                        "type": "integer",
                        "description": "글 조회수",
                        "example": 3
                      },
                      "board_writetime":
                      {
                        "type": "string",
                        "description": "글 작성시간",
                        "example": "2017-12-01 08:03:38"
                      },
                      "user_nickname":
                      {
                        "type": "string",
                        "description": "작성자 닉네임",
                        "example": "테스트"
                      }
                    }
                  }
                },
                "msg":
                {
                  "type": "string",
                  "description": "글 리스트 로딩 성공 메세지",
                  "example": "successful get board lists"
                }
              }
            }
          },
          "500":
          {
            "description": "get connection error // get board lists error",
            "schema":
            {
              "$ref": "#/definitions/failMessage"
            }
          }
        }
      }
    },
    "/board/content/:boardid/:useremail":
    {
      "get":
      {
        "tags": [
          "board"
        ],
        "summary": "글 내용 조회",
        "description": "글 상세 내용을 조회합니다. 글 조회시 글의 조회수가 1 증가합니다. Request : application/x-www-form-urlencoded",
        "operationId": "basicInfo",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
        {
          "in": "query",
          "name": "boardid",
          "description": "조회할 글의 ID",
          "required": true,
          "type": "integer"
        },
        {
          "in": "query",
          "name": "useremail",
          "description": "현재 글을 조회하는 유저의 이메일",
          "required": true,
          "type": "string"
        }],
        "responses":
        {
          "200":
          {
            "description": "글 상세정보 조회 성공",
            "schema":
            {
              "type": "object",
              "properties":
              {
                "status":
                {
                  "type": "string",
                  "description": "성공 알림",
                  "example": "success"
                },
                "data":
                {
                  "type": "object",
                  "properties":
                  {
                    "content":
                    {
                      "type": "object",
                      "properties":
                      {
                        "board_id":
                        {
                          "type": "integer",
                          "description": "조회하는 글의 id",
                          "example": 1
                        },
                        "board_title":
                        {
                          "type": "string",
                          "description": "글의 제목",
                          "example": "글쓰기테스트"
                        },
                        "board_content":
                        {
                          "type": "string",
                          "description": "글의 내",
                          "example": "글쓰기테스트입니다"
                        },
                        "board_image":
                        {
                          "type": "string",
                          "description": "글의 이미지 url",
                          "example": "image url"
                        },
                        "board_hits":
                        {
                          "type": "integer",
                          "description": "글의 조회수, 조회시 1 증가",
                          "example": 4
                        },
                        "board_writetime":
                        {
                          "type": "string",
                          "description": "글의 작성시간",
                          "example": "2017-12-01 08:03:38"
                        },
                        "board_bookmarked":
                        {
                          "type": "integer",
                          "description": "조회하는 유저가 이 글을 북마크 했는지 여부. 0 : 북마크하지 않음, 1 : 북마크 함",
                          "example": 0
                        },
                        "user_nickname":
                        {
                          "type": "string",
                          "description": "글의 작성자 닉네임",
                          "example": "테스트"
                        }
                      }
                    },
                    "comment":
                    {
                      "type": "array",
                      "items":
                      {
                        "type": "object",
                        "properties":
                        {
                          "comment_id":
                          {
                            "type": "integer",
                            "description": "댓글의 ID",
                            "example": 1
                          },
                          "comment_content":
                          {
                            "type": "string",
                            "description": "댓글 내용",
                            "example": "댓글테스트입니다"
                          },
                          "comment_writetime":
                          {
                            "type": "string",
                            "description": "댓글 작성시간",
                            "example": "2017-12-01 08:25:12"
                          },
                          "user_nickname":
                          {
                            "type": "string",
                            "description": "댓글 작성자 닉네임",
                            "example": "테스트"
                          }
                        }
                      }
                    }
                  }
                },
                "msg":
                {
                  "type": "string",
                  "description": "글 상세정보 로딩 성공 메세지",
                  "example": "successful load board content"
                }
              }
            }
          },
          "500":
          {
            "description": "get connection error // get board content error // get board comments error // get bookmark data error // update board hit error",
            "schema":
            {
              "$ref": "#/definitions/failMessage"
            }
          }
        }
      }
    },
    "/board/new/content":
    {
      "post":
      {
        "tags": [
          "board"
        ],
        "summary": "글 등록",
        "description": "게시판에 글을 등록합니다.. Request : multipart/form-data",
        "operationId": "basicInfo",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
        {
          "in": "body",
          "name": "body",
          "description": "글을 등록하기 위해 제목, 내용, 이미지, 작성자이메일 전달",
          "required": true,
          "schema":
          {
            "type": "object",
            "properties":
            {
              "title":
              {
                "description": "글의 제목",
                "type": "string",
                "minLength": 1,
                "required": true,
                "example": "글작성 테스트"
              },
              "content":
              {
                "description": "글의 내용",
                "type": "string",
                "minLength": 1,
                "required": true,
                "example": "글작성 테스트입니다. "
              },
              "image":
              {
                "description": "글에 들어갈 이미지파일. 필수 아님.",
                "required": false,
                "type": "image"
              },
              "email":
              {
                "description": "작성자 이메일",
                "type": "string",
                "minLength": 1,
                "required": true,
                "example": "test@sopt.org"
              }
            }
          }
        }],
        "responses":
        {
          "201":
          {
            "description": "글 등록 성공",
            "schema":
            {
              "type": "object",
              "properties":
              {
                "status":
                {
                  "type": "string",
                  "description": "성공 알림",
                  "example": "success"
                },
                "msg":
                {
                  "type": "string",
                  "description": "글 작성 성공 알림 메세지",
                  "example": "successful regist board data"
                }
              }
            }
          },
          "500":
          {
            "description": "get connection error // regist board data error",
            "schema":
            {
              "$ref": "#/definitions/failMessage"
            }
          }
        }
      }
    },
    "/board/new/comment":
    {
      "post":
      {
        "tags": [
          "board"
        ],
        "summary": "댓글 등록",
        "description": "글에 댓글을 등록합니다.. Request : application/x-www-form-urlencoded",
        "operationId": "basicInfo",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
        {
          "in": "body",
          "name": "body",
          "description": "댓글을 등록하기 위해 내용, 등록될 글의 id, 작성자이메일 전달",
          "required": true,
          "schema":
          {
            "type": "object",
            "properties":
            {
              "content":
              {
                "description": "댓글의 내용",
                "type": "string",
                "minLength": 1,
                "required": true,
                "example": "댓글작성 테스트입니다. "
              },
              "id":
              {
                "description": "댓글이 작성될 글의 id",
                "type": "integer",
                "minLength": 1,
                "required": true,
                "example": 1
              },
              "email":
              {
                "description": "작성자 이메일",
                "type": "string",
                "minLength": 1,
                "required": true,
                "example": "test@sopt.org"
              }
            }
          }
        }],
        "responses":
        {
          "201":
          {
            "description": "댓글 등록 성공",
            "schema":
            {
              "type": "object",
              "properties":
              {
                "status":
                {
                  "type": "string",
                  "description": "성공 알림",
                  "example": "success"
                },
                "msg":
                {
                  "type": "string",
                  "description": "글 작성 성공 알림 메세지",
                  "example": "successful regist comment data"
                }
              }
            }
          },
          "500":
          {
            "description": "get connection error // regist comment data error",
            "schema":
            {
              "$ref": "#/definitions/failMessage"
            }
          }
        }
      }
    },
    "/board/bookmark":
    {
      "put":
      {
        "tags": [
          "board"
        ],
        "summary": "북마크 등록/삭제",
        "description": "북마크를 등록하거나 삭제합니다. Request : application/x-www-form-urlencoded",
        "operationId": "basicInfo",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
        {
          "in": "body",
          "name": "body",
          "description": "북마크를 하기 위해 사용자의 email, 글의 id 전달",
          "required": true,
          "schema":
          {
            "type": "object",
            "properties":
            {
              "useremail":
              {
                "description": "북마크를 하는 유저의 이메일",
                "type": "string",
                "minLength": 1,
                "required": true,
                "example": "test@sopt.org"
              },
              "boardid":
              {
                "description": "북마크 추가/삭제 될 글의 id",
                "type": "integer",
                "minLength": 1,
                "required": true,
                "example": 1
              }
            }
          }
        }],
        "responses":
        {
          "201":
          {
            "description": "글 리스트",
            "schema":
            {
              "type": "object",
              "properties":
              {
                "status":
                {
                  "type": "string",
                  "description": "성공 알림",
                  "example": "success"
                },
                "msg":
                {
                  "type": "string",
                  "description": "북마크 추가/제거 성공 메세지",
                  "example": "successful regist bookmark // successful delete bookmark"
                }
              }
            }
          },
          "500":
          {
            "description": "get connection error // get bookmark data error // insert bookmark data error // delete bookmark data error",
            "schema":
            {
              "$ref": "#/definitions/failMessage"
            }
          }
        }
      }
    },
    "/mypage/bookmark/:useremail":
    {
      "get":
      {
        "tags": [
          "mypage"
        ],
        "summary": "북마크 글 리스트 조회",
        "description": "북마크한 글 리스트를 조회합니다. Request : application/x-www-form-urlencoded",
        "operationId": "basicInfo",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
        {
          "in": "query",
          "name": "useremail",
          "description": "북마크한 글을 조회하는 유저의 이메일",
          "required": true,
          "type": "string"
        }],
        "responses":
        {
          "200":
          {
            "description": "글 리스트",
            "schema":
            {
              "type": "object",
              "properties":
              {
                "status":
                {
                  "type": "string",
                  "description": "성공 알림",
                  "example": "success"
                },
                "data":
                {
                  "type": "array",
                  "items":
                  {
                    "type": "object",
                    "properties":
                    {
                      "board_id":
                      {
                        "type": "integer",
                        "description": "글의 ID",
                        "example": 1
                      },
                      "board_title":
                      {
                        "type": "string",
                        "description": "글 제목",
                        "example": "글쓰기 테스트"
                      },
                      "board_image":
                      {
                        "type": "string",
                        "description": "글 이미지 URL",
                        "example": "image url"
                      },
                      "board_hits":
                      {
                        "type": "integer",
                        "description": "글 조회수",
                        "example": 3
                      },
                      "board_writetime":
                      {
                        "type": "string",
                        "description": "글 작성시간",
                        "example": "2017-12-01 08:03:38"
                      },
                      "user_nickname":
                      {
                        "type": "string",
                        "description": "작성자 닉네임",
                        "example": "테스트"
                      }
                    }
                  }
                },
                "msg":
                {
                  "type": "string",
                  "description": "글 리스트 로딩 성공 메세지",
                  "example": "successful get bookmark lists"
                }
              }
            }
          },
          "500":
          {
            "description": "get connection error // get marked board data error",
            "schema":
            {
              "$ref": "#/definitions/failMessage"
            }
          }
        }
      }
    }
  },
  "definitions":
  {
    "failMessage":
    {
      "type": "object",
      "properties":
      {
        "status":
        {
          "type": "string",
          "example": "fail"
        },
        "message":
        {
          "type": "string",
          "example": "fail reason"
        }
      }
    }
  }
}